fst/src/edu/cmu/sphinx/fst/operations/ArcSort.java
/**
 * 
 * Copyright 1999-2012 Carnegie Mellon University.  
 * Portions Copyright 2002 Sun Microsystems, Inc.  
 * Portions Copyright 2002 Mitsubishi Electric Research Laboratories.
 * All Rights Reserved.  Use is subject to license terms.
 * 
 * See the file "license.terms" for information on usage and
 * redistribution of this file, and for a DISCLAIMER OF ALL 
 * WARRANTIES.
 *
 */

package edu.cmu.sphinx.fst.operations;

import java.util.Comparator;


import edu.cmu.sphinx.fst.arc.Arc;
import edu.cmu.sphinx.fst.fst.Fst;
import edu.cmu.sphinx.fst.state.State;

/**
 * @author John Salatas <jsalatas@users.sourceforge.net>
 *
 */
public class ArcSort {
	private ArcSort() {}
	
	public static <T extends Comparable<T>> void apply(Fst<T> fst, Comparator<Arc<T>> cmp) {
		for(int i=0;i<fst.getNumStates(); i++) {
			State<T> s = fst.getStateByIndex(i);
			s.arcSort(cmp);
		}
	}



}
